pyproject.toml
[project]
name = "chat-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]==0.115.12",
    "langchain==0.0.139",
    "openai==1.77.0",
    "pandas==2.2.3",
    "pypdf2==3.0.1",
    "python-dotenv==1.1.0",
    "redis==6.0.0",
    "tiktoken==0.9.0",
    "uvloop==0.21.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]
